Update repo for optimized image building (#3)
This updates the repo to work with the version of the Native Docker
Build System introduced in
https://github.com/sematic-ai/sematic/pull/874.

Specifically, it removes the explicit dependency on `sematic` in the
requirements file, as this is now being installed automatically if
missing. A token dependency is left in order to demonstrate the usage of
a requirements file.